clean up service store tests
