--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/advanced/pathContent.json
@@ -1,4 +1,0 @@
-{
-  "/trades/allTradesService" : "s_tradeId,s_traderDetails,s_tradeDetails\n1,abc:F_Name_1:L_Name_1,30:100\n2,abc:F_Name_1:L_Name_1,31:200\n3,abc:F_Name_2:L_Name_2,32:300\n4,abc:F_Name_2:L_Name_2,33:400\n5,abc:F_Name_1:L_Name_1,34:100\n6,abc:F_Name_1:L_Name_1,35:200\n7,abc:F_Name_2:L_Name_2,36:300\n8,abc:F_Name_2:L_Name_2,37:400\n9,abc:F_Name_2:L_Name_2,38:300\n10,abc:F_Name_2:L_Name_2,39:400",
-  "/products/getProductsByIdWithQuantity?idWithQuantity=30:100&idWithQuantity=31:200&idWithQuantity=32:300&idWithQuantity=33:400&idWithQuantity=34:100&idWithQuantity=35:200&idWithQuantity=36:300&idWithQuantity=37:400&idWithQuantity=38:300&idWithQuantity=39:400" : "s_productId,s_productName,s_description,s_quantity\n30,Product 30,Product 30 description,100\n31,Product 31,Product 31 description,200\n32,Product 32,Product 32 description,300\n33,Product 33,Product 33 description,400\n34,Product 34,Product 34 description,100\n35,Product 35,Product 35 description,200\n36,Product 36,Product 36 description,300\n37,Product 37,Product 37 description,400\n38,Product 38,Product 38 description,300\n39,Product 39,Product 39 description,400"
-}
