+++ legend-engine-executionPlan-execution-external-store-service/src/test/resources/features/batching/mappings/stub.json
@@ -1,0 +1,59 @@
+{
+  "mappings": [
+    {
+      "request": {
+        "method": "GET",
+        "url": "/trades/allTradesService"
+      },
+      "response": {
+        "status": 200,
+        "body": "s_tradeId,s_traderDetails,s_tradeDetails\n1,abc:F_Name_1:L_Name_1,30:100\n2,abc:F_Name_1:L_Name_1,31:200\n3,abc:F_Name_2:L_Name_2,32:300\n4,abc:F_Name_2:L_Name_2,33:400\n5,abc:F_Name_1:L_Name_1,34:100\n6,abc:F_Name_1:L_Name_1,35:200\n7,abc:F_Name_2:L_Name_2,36:300\n8,abc:F_Name_2:L_Name_2,37:400\n9,abc:F_Name_2:L_Name_2,38:300\n10,abc:F_Name_2:L_Name_2,39:400\n11,abc:F_Name_1:L_Name_1,35:200"
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "urlPathPattern": "/products/getProductsById",
+        "queryParameters": {
+          "id": {
+            "equalTo": "30,39,38,37,36,35,34,33,32,31"
+          }
+        }
+      },
+      "response": {
+        "status": 200,
+        "body": "s_productId,s_productName,s_description\n30,Product 30,Product 30 description\n31,Product 31,Product 31 description\n32,Product 32,Product 32 description\n33,Product 33,Product 33 description\n34,Product 34,Product 34 description\n35,Product 35,Product 35 description\n36,Product 36,Product 36 description\n37,Product 37,Product 37 description\n38,Product 38,Product 38 description\n39,Product 39,Product 39 description\n40,Product 40,Product 40 description"
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "urlPathPattern": "/products/getProductsById",
+        "queryParameters": {
+          "id": {
+            "equalTo": "34:100,31:200,37:400,32:300,36:300,38:300,39:400,30:100,35:200,33:400"
+          }
+        }
+      },
+      "response": {
+        "status": 200,
+        "body": "s_productId,s_productName,s_description\n30:100,Product 30,Product 30 description\n31:200,Product 31,Product 31 description\n32:300,Product 32,Product 32 description\n33:400,Product 33,Product 33 description\n34:100,Product 34,Product 34 description\n35:200,Product 35,Product 35 description\n36:300,Product 36,Product 36 description\n37:400,Product 37,Product 37 description\n38:300,Product 38,Product 38 description\n39:400,Product 39,Product 39 description"
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "urlPathPattern": "/products/getProductsByIdWithQuantity",
+        "queryParameters": {
+          "idWithQuantity": {
+            "equalTo": "31:200,39:400,35:200,38:300,37:400,33:400,30:100,32:300,34:100,36:300"
+          }
+        }
+      },
+      "response": {
+        "status": 200,
+        "body": "s_productId,s_productName,s_description,s_quantity\n30,Product 30,Product 30 description,100\n31,Product 31,Product 31 description,200\n32,Product 32,Product 32 description,300\n33,Product 33,Product 33 description,400\n34,Product 34,Product 34 description,100\n35,Product 35,Product 35 description,200\n36,Product 36,Product 36 description,300\n37,Product 37,Product 37 description,400\n38,Product 38,Product 38 description,300\n39,Product 39,Product 39 description,400"
+      }
+    }
+  ]
+}
