--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/fragment/mapping.pure
@@ -1,53 +1,0 @@
-// Copyright 2022 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-###Mapping
-Mapping meta::external::store::service::showcase::mapping::ServiceStoreMapping
-(
-    *meta::external::store::service::showcase::domain::Person[person_set]: ServiceStore
-    {
-        ~service [meta::external::store::service::showcase::store::EmployeesServiceStore] EmployeesService
-        (
-            ~path $service.response.employees
-        )
-    }
-
-    *meta::external::store::service::showcase::domain::Firm[firm_set]: ServiceStore
-    {
-        ~service [meta::external::store::service::showcase::store::EmployeesServiceStore] EmployeesService
-        (
-            ~path $service.response.firms
-        )
-    }
-
-    meta::external::store::service::showcase::domain::Person[person_set2]: ServiceStore
-    {
-        +firmId : Integer[1];
-
-        ~service [meta::external::store::service::showcase::store::EmployeesServiceStore] EmployeesServiceByFirmId
-        (
-            ~path $service.response.employees
-
-            ~paramMapping
-            (
-                firmId : $this.firmId
-            )
-        )
-    }
-
-    *meta::external::store::service::showcase::domain::Employment: XStore
-    {
-        employees[firm_set, person_set2] : $this.firmId == $that.firmId
-    }
-)
