--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/fragment/model.pure
@@ -1,53 +1,0 @@
-// Copyright 2022 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-###Pure
-import meta::external::store::service::showcase::domain::*;
-
-Class meta::external::store::service::showcase::domain::ApiResponse
-{
-    metadata  : Metadata[1];
-    employees : Person[*];
-    firms     : Firm[*];
-}
-
-Class meta::external::store::service::showcase::domain::Metadata
-{
-    noOfRecords : Integer[1];
-}
-
-Class meta::external::store::service::showcase::domain::Person
-{
-    firstName  : String[1];
-    lastName   : String[1];
-    middleName : String[0..1];
-}
-
-Class meta::external::store::service::showcase::domain::Firm
-{
-    firmName : String[1];
-    firmId   : Integer[1];
-    address  : Address[*];
-}
-
-Class meta::external::store::service::showcase::domain::Address
-{
-    street : String[1];
-}
-
-Association meta::external::store::service::showcase::domain::Employment
-{
-    employees : Person[*];
-    firm      : Firm[0..1];
-}
