--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/fragment/pathContent.json
@@ -1,6 +1,0 @@
-{
-  "/employees" : "{\n  \"metadata\": {\n    \"noOfRecords\": 3\n  },\n  \"employees\": [\n    {\n      \"firstName\": \"Nancy\",\n      \"lastName\": \"Fraher\",\n      \"dateOfBirth\": \"1970-12-13\",\n      \"firmId\": 1,\n      \"isAlive\": false,\n      \"heightInMeters\": 1.71\n    },\n    {\n      \"firstName\": \"Jason\",\n      \"lastName\": \"Schlichting\",\n      \"dateOfBirth\": \"1968-03-05\",\n      \"isAlive\": true,\n      \"firmId\": 2,\n      \"heightInMeters\": 1.82\n    },\n    {\n      \"firstName\": \"Jason2\",\n      \"lastName\": \"Schlichting\",\n      \"dateOfBirth\": \"1968-03-05\",\n      \"isAlive\": true,\n      \"firmId\": 3,\n      \"heightInMeters\": 1.82\n    }\n  ],\n  \"firms\": [\n    {\n      \"firmName\": \"Firm A\",\n      \"firmId\": 1,\n      \"address\": [\n        {\n          \"street\": \"Firm A address\"\n        }\n      ]\n    },\n    {\n      \"firmName\": \"Firm B\",\n      \"firmId\": 2,\n      \"address\": [\n        {\n          \"street\": \"Firm B address 1\"\n        },\n        {\n          \"street\": \"Firm B address 2\"\n        }\n      ]\n    },\n    {\n      \"firmName\": \"Firm c\",\n      \"firmId\": 3,\n      \"address\": []\n    }\n  ]\n}\n",
-  "/employeesByFirmId/1" : "{\n  \"metadata\": {\n    \"noOfRecords\": 1\n  },\n  \"employees\": [\n    {\n      \"firstName\": \"Nancy\",\n      \"lastName\": \"Fraher\",\n      \"dateOfBirth\": \"1970-12-13\",\n      \"firmId\": 1,\n      \"isAlive\": false,\n      \"heightInMeters\": 1.71\n    }\n  ],\n  \"firms\": [\n    {\n      \"firmName\": \"Firm A\",\n      \"firmId\": 1,\n      \"address\": [\n        {\n          \"street\": \"Firm A address\"\n        }\n      ]\n    }\n  ]\n}\n",
-  "/employeesByFirmId/2" : "{\n  \"metadata\": {\n    \"noOfRecords\": 1\n  },\n  \"employees\": [\n    {\n      \"firstName\": \"Jason\",\n      \"lastName\": \"Schlichting\",\n      \"dateOfBirth\": \"1968-03-05\",\n      \"isAlive\": true,\n      \"firmId\": 2,\n      \"heightInMeters\": 1.82\n    }\n ],\n  \"firms\": [\n    {\n      \"firmName\": \"Firm B\",\n      \"firmId\": 2,\n      \"address\": [\n        {\n          \"street\": \"Firm B address 1\"\n        },\n        {\n          \"street\": \"Firm B address 2\"\n        }\n      ]\n    }\n  ]\n}\n",
-  "/employeesByFirmId/3" : "{\n  \"metadata\": {\n    \"noOfRecords\": 1\n  },\n  \"employees\": [\n    {\n      \"firstName\": \"Jason2\",\n      \"lastName\": \"Schlichting\",\n      \"dateOfBirth\": \"1968-03-05\",\n      \"isAlive\": true,\n      \"firmId\": 3,\n      \"heightInMeters\": 1.82\n    }\n  ],\n  \"firms\": [\n    {\n      \"firmName\": \"Firm c\",\n      \"firmId\": 3,\n      \"address\": []\n    }\n  ]\n}\n"
-}
