--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/fragment/serviceStore.pure
@@ -1,68 +1,0 @@
-// Copyright 2022 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-###ServiceStore
-ServiceStore meta::external::store::service::showcase::store::EmployeesServiceStore
-(
-   description : 'Showcase Service Store Fragment support';
-
-   Service EmployeesService
-   (
-      path     : '/employees';
-      method   : GET;
-      security : [];
-      response : [meta::external::store::service::showcase::domain::ApiResponse <- meta::external::store::service::showcase::store::ApiResponseSchemaBinding];
-   )
-
-   Service EmployeesServiceByFirmId
-   (
-      path     : '/employeesByFirmId/{firmId}';
-      method   : GET;
-      parameters :
-      (
-         firmId : Integer (location = path)
-      );
-      security : [];
-      response : [meta::external::store::service::showcase::domain::ApiResponse <- meta::external::store::service::showcase::store::ApiResponseSchemaBinding];
-   )
-)
-
-###ExternalFormat
-Binding meta::external::store::service::showcase::store::ApiResponseSchemaBinding
-{
-  contentType   : 'application/json';
-  modelIncludes : [
-                    meta::external::store::service::showcase::domain::Metadata,
-                    meta::external::store::service::showcase::domain::Person,
-                    meta::external::store::service::showcase::domain::Firm,
-                    meta::external::store::service::showcase::domain::Address,
-                    meta::external::store::service::showcase::domain::ApiResponse
-                  ];
-}
-
-###Runtime
-Runtime meta::external::store::service::showcase::runtime::ServiceStoreRuntime
-{
-    mappings :
-    [
-        meta::external::store::service::showcase::mapping::ServiceStoreMapping
-    ];
-    connections :
-    [
-        meta::external::store::service::showcase::store::EmployeesServiceStore :
-        [
-            connection_1 : meta::external::store::service::showcase::connection::serviceStoreConnection
-        ]
-    ];
-}
