+++ legend-engine-executionPlan-execution-external-store-service/src/test/resources/showcase/json/mappings/stub.json
@@ -1,0 +1,181 @@
+{
+  "mappings": [
+    {
+      "request": {
+        "method": "GET",
+        "url": "/trades/allTradesService"
+      },
+      "response": {
+        "status": 200,
+        "jsonBody": [
+          {
+            "s_tradeId": "1",
+            "s_traderDetails": "abc:F_Name_1:L_Name_1",
+            "s_tradeDetails": "30:100"
+          },
+          {
+            "s_tradeId": "2",
+            "s_traderDetails": "abc:F_Name_1:L_Name_1",
+            "s_tradeDetails": "31:200"
+          },
+          {
+            "s_tradeId": "3",
+            "s_traderDetails": "abc:F_Name_2:L_Name_2",
+            "s_tradeDetails": "30:300"
+          },
+          {
+            "s_tradeId": "4",
+            "s_traderDetails": "abc:F_Name_2:L_Name_2",
+            "s_tradeDetails": "31:400"
+          }
+        ]
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "url": "/trades/1"
+      },
+      "response": {
+        "status": 200,
+        "jsonBody": [
+          {
+            "s_tradeId": "1",
+            "s_traderDetails": "abc:F_Name_1:L_Name_1",
+            "s_tradeDetails": "30:100"
+          }
+        ]
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "urlPath": "/trades/details",
+        "queryParameters": {
+          "trade.details": {
+            "equalTo": "30:100"
+          }
+        }
+      },
+      "response": {
+        "status": 200,
+        "jsonBody": [
+          {
+            "s_tradeId": "1",
+            "s_traderDetails": "abc:F_Name_1:L_Name_1",
+            "s_tradeDetails": "30:100"
+          }
+        ]
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "url": "/products/getProductById/30"
+      },
+      "response": {
+        "status": 200,
+        "jsonBody": {
+          "s_productId": "30",
+          "s_productName": "Product 30",
+          "s_description": "Product 30 description",
+          "s_synonyms": [
+            {
+              "s_name": "product 30 synonym 1",
+              "s_type": "isin"
+            },
+            {
+              "s_name": "product 30 synonym 2",
+              "s_type": "cusip"
+            }
+          ]
+        }
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "url": "/products/getProductById/31"
+      },
+      "response": {
+        "status": 200,
+        "jsonBody": {
+          "s_productId": "31",
+          "s_productName": "Product 31",
+          "s_description": "Product 31 description",
+          "s_synonyms": [
+            {
+              "s_name": "product 31 synonym 1",
+              "s_type": "isin"
+            },
+            {
+              "s_name": "product 31 synonym 2",
+              "s_type": "cusip"
+            }
+          ]
+        }
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "urlPath": "/products/getProductByName",
+        "queryParameters": {
+          "name": {
+            "equalTo": "product 30"
+          }
+        }
+      },
+      "response": {
+        "status": 200,
+        "jsonBody": {
+          "s_productId": "30",
+          "s_productName": "Product 30",
+          "s_description": "Product 30 description",
+          "s_synonyms": [
+            {
+              "s_name": "product 30 synonym 1",
+              "s_type": "isin"
+            },
+            {
+              "s_name": "product 30 synonym 2",
+              "s_type": "cusip"
+            }
+          ]
+        }
+      }
+    },
+    {
+      "request": {
+        "method": "GET",
+        "urlPath": "/products/getProductByNameAndDescription",
+        "queryParameters": {
+          "name": {
+            "equalTo": "product 30"
+          },
+          "description": {
+            "equalTo": "product 30 description"
+          }
+        }
+      },
+      "response": {
+        "status": 200,
+        "jsonBody": {
+          "s_productId": "30",
+          "s_productName": "Product 30",
+          "s_description": "Product 30 description",
+          "s_synonyms": [
+            {
+              "s_name": "product 30 synonym 1",
+              "s_type": "isin"
+            },
+            {
+              "s_name": "product 30 synonym 2",
+              "s_type": "cusip"
+            }
+          ]
+        }
+      }
+    }
+  ]
+}
