--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/showcase/json/pathContent.json
@@ -1,9 +1,0 @@
-{
-  "/trades/allTradesService" : "[\n  {\n    \"s_tradeId\": \"1\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"30:100\"\n  },\n  {\n    \"s_tradeId\": \"2\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"31:200\"\n  },\n  {\n    \"s_tradeId\": \"3\",\n    \"s_traderDetails\": \"abc:F_Name_2:L_Name_2\",\n    \"s_tradeDetails\": \"30:300\"\n  },\n  {\n    \"s_tradeId\": \"4\",\n    \"s_traderDetails\": \"abc:F_Name_2:L_Name_2\",\n    \"s_tradeDetails\": \"31:400\"\n  }\n]",
-  "/trades/1" : "[\n  {\n    \"s_tradeId\": \"1\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"30:100\"\n  }\n]",
-  "/trades/details?trade.details=30:100" : "[\n  {\n    \"s_tradeId\": \"1\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"30:100\"\n  }\n]",
-  "/products/getProductById/30" : "{\n    \"s_productId\": \"30\",\n    \"s_productName\": \"Product 30\",\n    \"s_description\": \"Product 30 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 30 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 30 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }\n",
-  "/products/getProductById/31" : "{\n    \"s_productId\": \"31\",\n    \"s_productName\": \"Product 31\",\n    \"s_description\": \"Product 31 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 31 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 31 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }\n",
-  "/products/getProductByName?name=product 30" : "{\n    \"s_productId\": \"30\",\n    \"s_productName\": \"Product 30\",\n    \"s_description\": \"Product 30 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 30 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 30 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }\n",
-  "/products/getProductByNameAndDescription?name=product 30&description=product 30 description" : "{\n    \"s_productId\": \"30\",\n    \"s_productName\": \"Product 30\",\n    \"s_description\": \"Product 30 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 30 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 30 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }\n"
-}
