--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/union/pathContent.json
@@ -1,11 +1,0 @@
-{
-  "/trades/allTradesService1" : "[\n  {\n    \"s_tradeId\": \"1\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"30:100\"\n  },\n  {\n    \"s_tradeId\": \"2\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"31:200\"\n  },\n  {\n    \"s_tradeId\": \"3\",\n    \"s_traderDetails\": \"abc:F_Name_2:L_Name_2\",\n    \"s_tradeDetails\": \"30:300\"\n  },\n  {\n    \"s_tradeId\": \"4\",\n    \"s_traderDetails\": \"abc:F_Name_2:L_Name_2\",\n    \"s_tradeDetails\": \"31:400\"\n  }\n]",
-  "/trades/allTradesService2" : "[\n  {\n    \"s_tradeId\": \"5\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"40:100\"\n  },\n  {\n    \"s_tradeId\": \"6\",\n    \"s_traderDetails\": \"abc:F_Name_1:L_Name_1\",\n    \"s_tradeDetails\": \"41:200\"\n  },\n  {\n    \"s_tradeId\": \"7\",\n    \"s_traderDetails\": \"abc:F_Name_2:L_Name_2\",\n    \"s_tradeDetails\": \"40:300\"\n  },\n  {\n    \"s_tradeId\": \"8\",\n    \"s_traderDetails\": \"abc:F_Name_2:L_Name_2\",\n    \"s_tradeDetails\": \"41:400\"\n  }\n]",
-  "/products/getAllProducts1" : "[\n  {\n    \"s_productId\": \"30\",\n    \"s_productName\": \"Product 30\",\n    \"s_description\": \"Product 30 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 30 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 30 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  },\n  {\n    \"s_productId\": \"31\",\n    \"s_productName\": \"Product 31\",\n    \"s_description\": \"Product 31 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 31 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 31 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }\n]",
-  "/products/getAllProducts2" : "[\n  {\n    \"s_productId\": \"40\",\n    \"s_productName\": \"Product 40\",\n    \"s_description\": \"Product 40 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 40 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 40 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  },\n  {\n    \"s_productId\": \"41\",\n    \"s_productName\": \"Product 41\",\n    \"s_description\": \"Product 41 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 41 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 41 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }\n]",
-  "/products/getProductById1/31" : "[{\n    \"s_productId\": \"31\",\n    \"s_productName\": \"Product 31\",\n    \"s_description\": \"Product 31 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 31 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 31 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }]",
-  "/products/getProductById1/41" : "[]",
-  "/products/getProductById2/31" : "[]",
-  "/products/getProductById2/41" : "[{\n    \"s_productId\": \"41\",\n    \"s_productName\": \"Product 41\",\n    \"s_description\": \"Product 41 description\",\n    \"s_synonyms\": [\n      {\n        \"s_name\": \"product 41 synonym 1\",\n        \"s_type\": \"isin\"\n      },\n      {\n        \"s_name\": \"product 41 synonym 2\",\n        \"s_type\": \"cusip\"\n      }\n    ]\n  }]",
-  "/employees" : "[{\"firstName\":\"FirstName ServiceStore\", \"lastName\" : \"LastName ServiceStore\", \"firmId\":\"ServiceStore\"}]"
-}
