--- legend-engine-executionPlan-execution-external-store-service/src/test/resources/union/serviceStore.pure
@@ -1,158 +1,0 @@
-// Copyright 2021 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-###ServiceStore
-ServiceStore meta::external::store::service::showcase::store::TradeProductServiceStore
-(
-   description : 'Service Store Union Example';
-
-   ServiceGroup TradeServices
-   (
-      path : '/trades';
-
-      Service AllTradeService1
-      (
-         path : '/allTradesService1';
-         method : GET;
-         security : [];
-         response : [meta::external::store::service::showcase::domain::S_Trade <- meta::external::store::service::showcase::store::tradeServiceStoreSchemaBinding];
-      )
-
-      Service AllTradeService2
-      (
-         path : '/allTradesService2';
-         method : GET;
-         security : [];
-         response : [meta::external::store::service::showcase::domain::S_Trade <- meta::external::store::service::showcase::store::tradeServiceStoreSchemaBinding];
-      )
-   )
-
-   ServiceGroup ProductServices
-   (
-      path : '/products';
-
-      Service ProductByIdService1
-      (
-         path : '/getProductById1/{id}';
-         method : GET;
-         parameters :
-         (
-            id : String (location = path)
-         );
-         security : [];
-         response : [meta::external::store::service::showcase::domain::S_Product <- meta::external::store::service::showcase::store::prodServiceStoreSchemaBinding];
-      )
-
-      Service AllProductsService1
-      (
-         path : '/getAllProducts1';
-         method : GET;
-         security : [];
-         response : [meta::external::store::service::showcase::domain::S_Product <- meta::external::store::service::showcase::store::prodServiceStoreSchemaBinding];
-      )
-
-      Service AllProductsService2
-      (
-         path : '/getAllProducts2';
-         method : GET;
-         security : [];
-         response : [meta::external::store::service::showcase::domain::S_Product <- meta::external::store::service::showcase::store::prodServiceStoreSchemaBinding];
-      )
-
-      Service ProductByIdService2
-      (
-         path : '/getProductById2/{id}';
-         method : GET;
-         parameters :
-         (
-            id : String (location = path)
-         );
-         security : [];
-         response : [meta::external::store::service::showcase::domain::S_Product <- meta::external::store::service::showcase::store::prodServiceStoreSchemaBinding];
-      )
-   )
-)
-
-ServiceStore meta::external::store::service::showcase::store::EmployeeServiceStore
-(
-   description : 'Service Store Union Example';
-
-   Service EmployeeService
-   (
-      path : '/employees';
-      method : GET;
-      security : [];
-      response : [meta::external::store::service::showcase::domain::Person <- meta::external::store::service::showcase::store::employeesServiceStoreSchemaBinding];
-   )
-)
-
-###ExternalFormat
-Binding meta::external::store::service::showcase::store::tradeServiceStoreSchemaBinding
-{
-  contentType   : 'application/json';
-  modelIncludes : [ meta::external::store::service::showcase::domain::S_Trade ];
-}
-
-Binding meta::external::store::service::showcase::store::prodServiceStoreSchemaBinding
-{
-  contentType   : 'application/json';
-  modelIncludes : [ meta::external::store::service::showcase::domain::S_Product, meta::external::store::service::showcase::domain::S_Synonym ];
-}
-
-Binding meta::external::store::service::showcase::store::employeesServiceStoreSchemaBinding
-{
-  contentType   : 'application/json';
-  modelIncludes : [ meta::external::store::service::showcase::domain::Person ];
-}
-
-###Runtime
-Runtime meta::external::store::service::showcase::runtime::ServiceStoreRuntime
-{
-    mappings :
-    [
-        meta::external::store::service::showcase::mapping::ServiceStoreMapping
-    ];
-    connections :
-    [
-        meta::external::store::service::showcase::store::TradeProductServiceStore :
-        [
-            connection_1 : meta::external::store::service::showcase::connection::serviceStoreConnection
-        ],
-        meta::external::store::service::showcase::store::EmployeeServiceStore :
-        [
-            connection_2 : meta::external::store::service::showcase::connection::serviceStoreConnection2
-        ],
-        ModelStore :
-        [
-            connection_3 : meta::external::store::service::showcase::connection::modelChainConnection,
-            connection_4 : meta::external::store::service::showcase::connection::jsonModelConnection
-        ]
-    ];
-}
-
-###Connection
-ModelChainConnection meta::external::store::service::showcase::connection::modelChainConnection
-{
-    mappings :
-    [
-        meta::external::store::service::showcase::mapping::ServiceStoreMapping
-    ];
-}
-
-
-JsonModelConnection meta::external::store::service::showcase::connection::jsonModelConnection
-{
-    class : meta::external::store::service::showcase::domain::Person;
-    url   : 'data:application/json,{"firstName":"FirstName Model", "lastName" : "LastName Model", "firmId":"Model"}';
-}
